Fixed selection border after resizing

+++ src/main/com.calabi.pixelator/com/calabi/pixelator/control/image/OutlineRect.java
@@ -64,10 +64,10 @@ public class OutlineRect extends ShapeStack {
     }
 
     public void setEdges(int x1, int y1, int x2, int y2) {
-        this.x1.set(Math.min(getPixelWidth(), Math.max(0, x1 < x2 ? x1 : x2)));
-        this.y1.set(Math.min(getPixelHeight(), Math.max(0, y1 < y2 ? y1 : y2)));
-        this.x2.set(Math.min(getPixelWidth(), Math.max(0, (x1 < x2 ? x2 : x1) + 1)));
-        this.y2.set(Math.min(getPixelHeight(), Math.max(0, (y1 < y2 ? y2 : y1) + 1)));
+        this.x1.set(Math.min((int) getPixelWidth(), Math.max(0, x1 < x2 ? x1 : x2)));
+        this.y1.set(Math.min((int) getPixelHeight(), Math.max(0, y1 < y2 ? y1 : y2)));
+        this.x2.set(Math.min((int) getPixelWidth(), Math.max(0, (x1 < x2 ? x2 : x1) + 1)));
+        this.y2.set(Math.min((int) getPixelHeight(), Math.max(0, (y1 < y2 ? y2 : y1) + 1)));
     }
 
     public void clear() {
